requirements.txt
Django==2.2
django-cors-headers==3.7.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
psycopg2